+++ apps/consultation-portal/components/Card/SimpleCardSkeleton.tsx
@@ -5,6 +5,7 @@ export type SimpleCardSkeletonProps = {
   borderWidth?: 'large' | 'standard'
   borderRadius?: 'large' | 'standard'
   children: any
+  className?: string
 }
 
 export const SimpleCardSkeleton = ({
@@ -12,6 +13,7 @@ export const SimpleCardSkeleton = ({
   borderWidth,
   borderRadius,
   children,
+  className,
 }: SimpleCardSkeletonProps) => {
   return (
     <Box
@@ -21,6 +23,7 @@ export const SimpleCardSkeleton = ({
       borderRadius={borderRadius ? borderRadius : 'standard'}
       paddingX={4}
       paddingY={3}
+      className={className}
     >
       {children}
     </Box>
