+++ apps/consultation-portal/screens/Case/components/Stakeholders.css.ts
@@ -1,0 +1,13 @@
+import { style } from '@vanilla-extract/css'
+import { theme } from '@island.is/island-ui/theme'
+
+export const relativeBox = style({
+  position: 'relative',
+})
+
+export const blowout = style({
+  position: 'absolute',
+  top: 0,
+  right: 0,
+  margin: theme.spacing[2],
+})
